requirements.txt
flask==3.0.3
gunicorn==21.2.0
gevent==24.2.1
python-telegram-bot==20.6
openai>=1.12.0
stripe>=8.0.0
aiohttp==3.9.5
apscheduler==3.10.4
python-dotenv==1.0.1
supabase==2.3.4
requests>=2.31.0
